git-svn-id: file:///mnt/stora_backup/satt/backup-before-new-install/var/svn/repos/software/EasyLocalpp/trunk@127 1ceef749-5ca8-4b7e-9fa7-ff5341897c31

+++ src/helpers/NeighborhoodExplorer.hh
@@ -97,7 +97,7 @@ public:
   virtual void AddDeltaCostComponent(AbstractDeltaCostComponent<Input,State,Move,CFtype>& dcc);
   
   virtual unsigned int DeltaCostComponents()
-  { return delta_cost_component.size(); }
+  { return (unsigned int)delta_cost_component.size(); }
   
   virtual AbstractDeltaCostComponent<Input,State,Move,CFtype>& DeltaCostComponent(unsigned i)
   { return *delta_cost_component[i]; }
